requirements.txt
flask
requests
google-api-python-client
oauth2client
python-dotenv
pytrends
datetime